requirements.txt
pdfplumber>=0.11
python-docx>=1.0
reportlab>=4.0
